@@ -2,20 +2,20 @@
   "name": "game-of-life",
   "description": "React version of Conway's Game Of Life",
   "scripts": {
-    "start": "webpack-dev-server --config webpack.config.dev.js --port 3000",
-    "build:prod": "webpack --env.prod --config webpack.config.js --progress --color"
+    "start": "NODE_OPTIONS=--openssl-legacy-provider webpack-dev-server --config webpack.config.dev.js --port 3000",
+    "build:prod": "NODE_OPTIONS=--openssl-legacy-provider webpack --env.prod --config webpack.config.js --progress --color"
   },
   "dependencies": {
     "bootstrap": "4.3.1",
     "immer": "2.1.3",
     "prop-types": "15.7.2",
-    "react": "16.8.4",
+    "react": "^19.2.0",
     "react-cli": "^0.3.1",
-    "react-dom": "16.8.4",
-    "react-redux": "6.0.1",
+    "react-dom": "^19.2.0",
+    "react-redux": "^9.1.2",
     "react-router-dom": "5.0.0",
     "react-toastify": "4.5.2",
-    "redux": "4.0.1",
+    "redux": "^5.0.1",
     "redux-thunk": "2.3.0",
     "reselect": "4.0.0"
   },
@@ -26,8 +26,6 @@
     "babel-preset-react-app": "7.0.2",
     "css-loader": "2.1.1",
     "cssnano": "4.1.10",
-    "enzyme": "3.9.0",
-    "enzyme-adapter-react-16": "1.11.2",
     "eslint": "5.15.2",
     "eslint-loader": "2.1.2",
     "eslint-plugin-import": "2.16.0",
@@ -41,7 +39,7 @@
     "node-fetch": "^2.3.0",
     "npm-run-all": "4.1.5",
     "postcss-loader": "3.0.0",
-    "react-test-renderer": "16.8.4",
+    "react-test-renderer": "^19.2.0",
     "react-testing-library": "6.0.0",
     "redux-immutable-state-invariant": "2.1.0",
     "redux-mock-store": "1.5.3",
